+++ src/editors/starrating.js
@@ -10,8 +10,16 @@ JSONEditor.defaults.editors.starrating = JSONEditor.defaults.editors.string.exte
     this.ratingContainer = document.createElement('div');
     this.ratingContainer.classList.add('starrating');
 
-    this.enum_values = this.schema.enum;
-    this.radioGroup =[];
+    // Emulate the old "rating" editor parameters
+    if (this.schema.enum == undefined) {
+      var max = this.schema.maximum ? this.schema.maximum : 5;
+      if (this.schema.exclusiveMaximum) max--;
+      this.enum_values = [];
+      for (var k=0;k<max;k++) this.enum_values.push(k+1);
+    }
+    else this.enum_values = this.schema.enum;
+
+    this.radioGroup = [];
 
     var radioInputEventhandler = function(e) {
       e.preventDefault();
@@ -22,7 +30,7 @@ JSONEditor.defaults.editors.starrating = JSONEditor.defaults.editors.string.exte
 
     for(var i = this.enum_values.length-1; i>-1; i--) {
 
-      var id = this.key + '-' + i;
+      var id = this.formname + (i+1);
 
       // form radio elements
       var radioInput = this.theme.getFormInputField('radio');
@@ -54,7 +62,7 @@ JSONEditor.defaults.editors.starrating = JSONEditor.defaults.editors.string.exte
 
     if(this.schema.readOnly || this.schema.readonly) {
       this.always_disabled = true;
-      for (var j = 0; i<this.radioGroup.length; j++) {
+      for (var j = 0; j<this.radioGroup.length; j++) {
         this.radioGroup[j].disabled = true;
       }
       this.ratingContainer.classList.add('readonly');
@@ -64,9 +72,12 @@ JSONEditor.defaults.editors.starrating = JSONEditor.defaults.editors.string.exte
     ratingsContainerWrapper.appendChild(this.ratingContainer);
 
     this.input = ratingsContainerWrapper;
-    
+
     this.control = this.theme.getFormControl(this.label, ratingsContainerWrapper, this.description, this.infoButton);
     this.container.appendChild(this.control);
+
+    this.refreshValue();
+
   },
   enable: function() {
     if(!this.always_disabled) {
@@ -94,6 +105,15 @@ JSONEditor.defaults.editors.starrating = JSONEditor.defaults.editors.string.exte
   getNumColumns: function() {
     return 2;
   },
+  getValue: function() {
+    if (!this.dependenciesFulfilled) {
+      return undefined;
+    }
+    if (this.schema.type == 'integer') {
+      return this.value===''?undefined:this.value*1;
+    }
+    return this.value;
+  },
   setValue: function (val) {
     for(var i = 0; i < this.radioGroup.length; i++) {
 
@@ -103,7 +123,7 @@ JSONEditor.defaults.editors.starrating = JSONEditor.defaults.editors.string.exte
         if(this.options.displayValue) {
           this.displayRating.innerHTML = this.value;
         }
-        this.onChange();
+        this.onChange(true);
         break;
       }
     }